Moved static interface methods from NumberStore to UniversalNumberStore for Java SE6 compat.
